refactor: modularize runtime components and update utility imports
- Refactored the application structure by modularizing runtime components, including initial state management, layout configuration, and request handling.
- Updated utility imports to reflect new paths, enhancing code organization and maintainability.
- Introduced new constants for login paths and response statuses to streamline error handling and user authentication processes.
- Improved the rendering logic for various components by utilizing the newly structured utility functions for formatting and rendering data.
