+++ stockManager/front/src/pages/Data/components/AnalysisList.tsx
@@ -2,7 +2,8 @@ import { Col, Row, Statistic, Table, Tooltip } from 'antd';
 import React, { useMemo } from 'react';
 import type { ColumnsType } from 'antd/lib/table';
 import { useIsMobile } from '@/hooks/useIsMobile';
-import { colorFromValue, renderAmount as renderAmountTool } from '@/utils/renderTool';
+import { colorFromValue } from '@/utils/format/stock';
+import { renderAmount as renderAmountTool } from '@/utils/format/render';
 import { getHeaderStatisticStyles } from '@/utils/statisticStyles';
 import { useStockProfitModal } from '@/components/Common/StockProfitModal';
 import './index.less';
