+++ stockManager/front/src/pages/StockList/components/OverallBoard.tsx
@@ -1,57 +1,23 @@
-import { Statistic, Row, Col, InputNumber, Form, App, Divider } from 'antd';
+import { Statistic, Row, Col, Divider } from 'antd';
 import { useState } from 'react';
 import { UpOutlined, DownOutlined } from '@ant-design/icons';
-import { updateIncomeCash } from '@/services/api';
-import { colorFromValue } from '@/utils/renderTool';
+import { colorFromValue } from '@/utils/format/stock';
 import { useIsMobile } from '@/hooks/useIsMobile';
-import { RESPONSE_STATUS } from '@/utils/constants';
-import { useCashFlowModal } from '@/components/Common/CashFlowModal';
+import {
+  EXPANDED_STATISTICS,
+  MAIN_STATISTICS,
+  resolveOverallStat,
+  STAT_ACTIONS,
+  type OverallBoardActions,
+  type OverallStatConfig,
+} from './overallStat';
+import { useOverallBoardActions } from './useOverallBoardActions';
 import './index.less';
-
-// ==================== 类型定义 ====================
 
 type OverallBoardProps = {
   data: API.Overall;
   onModifySuccess?: () => void;
 };
-
-type StatisticItemConfig = {
-  key: keyof API.Overall;
-  title: string;
-  showColor?: boolean;
-  isMain?: boolean;
-  /** 自定义数值展示，默认两位小数 */
-  format?: (value: number) => string;
-  /** 返回 false 时不渲染该项 */
-  when?: (data: API.Overall) => boolean;
-};
-
-// ==================== 配置 ====================
-
-/** 主要统计项（始终显示） */
-const MAIN_STATISTICS: StatisticItemConfig[] = [
-  { key: 'offsetToday', title: '当日盈亏', showColor: true, isMain: true },
-  { key: 'totalAsset', title: '总资产', isMain: true },
-  { key: 'xirrAnnualized', title: 'XIRR年化', showColor: true, isMain: true },
-];
-
-/** 展开后的统计项 */
-const EXPANDED_STATISTICS: StatisticItemConfig[] = [
-  { key: 'offsetCurrent', title: '浮动盈亏', showColor: true },
-  { key: 'offsetTotal', title: '累计盈亏', showColor: true },
-  { key: 'totalValue', title: '市值', showColor: true },
-  { key: 'totalCash', title: '现金' },
-  { key: 'incomeCash', title: '其它现金收入' },
-  { key: 'originCash', title: '总入金' },
-  {
-    key: 'hkdCnyRate',
-    title: '港币汇率 (HKD/CNY)',
-    format: (v) => v.toFixed(4),
-    when: (d) => d.hkdCnyRate != null && d.hkdCnyRate > 0,
-  },
-];
-
-// ==================== 子组件 ====================
 
 type StatisticItemProps = {
   title: string;
@@ -72,7 +38,7 @@ const StatisticItem: React.FC<StatisticItemProps> = ({
 }) => {
   const isMobile = useIsMobile();
   const clickable = !!onClick;
-  
+
   const styles = {
     title: {
       fontSize: isMobile ? 12 : isMain ? 14 : 13,
@@ -94,101 +60,44 @@ const StatisticItem: React.FC<StatisticItemProps> = ({
   );
 };
 
-// ==================== 组件 ====================
+const renderStatistics = (
+  configs: OverallStatConfig[],
+  data: API.Overall,
+  actions: OverallBoardActions,
+) =>
+  configs
+    .filter(({ when }) => !when || when(data))
+    .map(({ key, title, showColor, isMain }) => {
+      const { value, numeric } = resolveOverallStat(key, data);
+      const actionKey = STAT_ACTIONS[key];
+      const onClick = actionKey ? actions[actionKey] : undefined;
+
+      return (
+        <StatisticItem
+          key={key}
+          title={title}
+          value={value}
+          numericValue={showColor ? numeric : undefined}
+          showColor={showColor}
+          isMain={isMain}
+          onClick={onClick}
+        />
+      );
+    });
 
 export const OverallBoard: React.FC<OverallBoardProps> = ({ data, onModifySuccess }) => {
-  const [form] = Form.useForm();
-  const { modal } = App.useApp();
   const [isExpanded, setIsExpanded] = useState(false);
-  const { showCashFlow } = useCashFlowModal();
   const isMobile = useIsMobile();
-
-  /** 可点击项的回调映射 */
-  const clickHandlers: Partial<Record<keyof API.Overall, () => void>> = {
-    incomeCash: () => {
-
-      const totalAsset = data.totalAsset || 0;
-      const incomeCash = data.incomeCash || 0;
-      form.setFieldsValue({ incomeCash: incomeCash, totalAsset: totalAsset });
-      
-      const fixedPart = totalAsset - incomeCash;
-
-      modal.confirm({
-        title: '编辑现金收入',
-        icon: null,
-        content: (
-          <Form form={form} layout="vertical" name="incomeCash" style={{ marginTop: 30 }}>
-            <Form.Item label="现金收入" name="incomeCash" rules={[{ required: true, message: '请输入现金收入' }]}>
-              <InputNumber
-                style={{ width: '100%' }}
-                precision={2}
-                step={0.01}
-                onChange={(v) => form.setFieldsValue({ totalAsset: fixedPart + ((v as number) || 0) })}
-              />
-            </Form.Item>
-            <Form.Item label="总资产" name="totalAsset" rules={[{ required: true, message: '请输入总资产' }]}>
-              <InputNumber
-                style={{ width: '100%' }}
-                precision={2}
-                step={0.01}
-                onChange={(v) => form.setFieldsValue({ incomeCash: ((v as number) || 0) - fixedPart })}
-              />
-            </Form.Item>
-          </Form>
-        ),
-        onOk: async () => {
-          const { incomeCash } = await form.validateFields();
-          const response = await updateIncomeCash(incomeCash);
-          if (response.status === RESPONSE_STATUS.SUCCESS) onModifySuccess?.();
-        },
-      });
-    },
-    originCash: () => showCashFlow({ totalCashIn: data.originCash || 0, cashFlowList: data.cashFlowList || [] }),
-  };
-
-  /** 格式化数值为字符串 */
-  const formatValue = (
-    rawValue: string | number | API.CashFlowRecord[] | undefined,
-    format?: (value: number) => string,
-  ): { value: string; numericValue: number } => {
-    if (typeof rawValue === 'string') {
-      const numericValue = parseFloat(rawValue.replace('%', '')) || 0;
-      return { value: rawValue, numericValue };
-    }
-    if (typeof rawValue === 'number') {
-      return {
-        value: format ? format(rawValue) : rawValue.toFixed(2),
-        numericValue: rawValue,
-      };
-    }
-    return { value: '0.00', numericValue: 0 };
-  };
-
-  /** 渲染统计项列表 */
-  const renderStatistics = (configs: StatisticItemConfig[]) =>
-    configs
-      .filter(({ when }) => !when || when(data))
-      .map(({ key, showColor, isMain, title, format }) => {
-        const { value, numericValue } = formatValue(data[key], format);
-
-        return (
-          <StatisticItem
-            key={key}
-            title={title}
-            value={value}
-            numericValue={showColor ? numericValue : undefined}
-            showColor={showColor}
-            isMain={isMain}
-            onClick={clickHandlers[key]}
-          />
-        );
-      });
+  const actions = useOverallBoardActions({ data, onModifySuccess });
 
   return (
     <div className="overall-board-wrapper">
-      <Row gutter={[16, 16]}>{renderStatistics(MAIN_STATISTICS)}</Row>
+      <Row gutter={[16, 16]}>{renderStatistics(MAIN_STATISTICS, data, actions)}</Row>
 
-      <div className={`expand-divider-wrapper ${isExpanded ? 'expanded' : 'collapsed'}`} onClick={() => setIsExpanded(!isExpanded)}>
+      <div
+        className={`expand-divider-wrapper ${isExpanded ? 'expanded' : 'collapsed'}`}
+        onClick={() => setIsExpanded(!isExpanded)}
+      >
         <Divider className="expand-divider" styles={{ root: { margin: 0 }, content: { margin: '0 8px' } }}>
           {isExpanded ? <UpOutlined className="expand-icon" /> : <DownOutlined className="expand-icon" />}
         </Divider>
@@ -196,7 +105,7 @@ export const OverallBoard: React.FC<OverallBoardProps> = ({ data, onModifySucces
 
       {isExpanded && (
         <Row gutter={[16, 16]} style={{ marginTop: isMobile ? 16 : 24 }}>
-          {renderStatistics(EXPANDED_STATISTICS)}
+          {renderStatistics(EXPANDED_STATISTICS, data, actions)}
         </Row>
       )}
     </div>
