+++ stockManager/front/src/runtime/bootstrap.ts
@@ -1,0 +1,7 @@
+// @ts-ignore
+import { version } from '../../package.json';
+
+console.log(
+  `%c Stock Manager v${version} `,
+  'background: #1890ff; color: #fff; padding: 4px 8px; border-radius: 4px; font-weight: bold;'
+);
