+++ stockManager/front/src/runtime/constants.ts
@@ -1,0 +1,1 @@
+export const LOGIN_PATH = '/login';
