+++ stockManager/front/src/utils/apiConstants.ts
@@ -1,6 +1,3 @@
-/** 移动端断点（与 styles/variables.less @screen-md 一致） */
-export const SCREEN_MD_MAX = 768;
-
 /**
  * 状态码常量定义
  * 与后端 constants.py 保持一致
@@ -47,13 +44,3 @@ const HTTP_STATUS_CONFIG = {
 export const HTTP_CODE_MESSAGE: Record<number, string> = Object.fromEntries(
   Object.values(HTTP_STATUS_CONFIG).map(({ code, message }) => [code, message])
 );
-
-/**
- * 环境标签颜色配置
- */
-export const ENV_TAG_COLORS = {
-  dev: 'orange',
-  test: 'green',
-  pre: '#87d068',
-} as const;
-
