+++ stockManager/front/src/utils/layoutConstants.ts
@@ -1,0 +1,11 @@
+/** 移动端断点（与 styles/variables.less @screen-md 一致） */
+export const SCREEN_MD_MAX = 768;
+
+/**
+ * 环境标签颜色配置
+ */
+export const ENV_TAG_COLORS = {
+  dev: 'orange',
+  test: 'green',
+  pre: '#87d068',
+} as const;
